Support weighting in confidence and significance
