+++ neleval/configs.py
@@ -153,6 +153,9 @@ def _expand(measures):
 
 
 def parse_measures(in_measures, incl_clustering=True, allow_unknown=False):
+    """Expand measure lists and validate, but do not map strings to Measures
+
+    """
     # flatten nested sequences and expand group names
     measures = _expand(in_measures)
     # remove duplicates while maintaining order
@@ -192,7 +195,7 @@ def get_measure(name, weighting=None):
         measure = Measure.from_string(name)
     else:
         measure = MEASURES[name]
-    return measure.get_weighted(weighting)
+    return measure.with_weighting(weighting)
 
 
 def get_measure_choices():
@@ -268,3 +271,27 @@ class ListMeasures(object):
                        metavar='NAME', help=MEASURE_HELP)
         p.set_defaults(cls=cls)
         return p
+
+
+class TypeWeighting:
+    def __init__(self, path):
+        self.values = {}
+        with open(path) as f:
+            for l in f:
+                gold, sys, weight = l.split('\t')
+                weight = float(weight)
+                self.values[gold, sys] = weight
+
+    def __call__(self, gold_type, sys_type):
+        if gold_type == sys_type:
+            return 1
+        return self.values.get((gold_type, sys_type), 0)
+
+
+def load_weighting(type_weights=None):
+    """type_weights is a path
+    """
+    weighting = {}
+    if type_weights is not None:
+        weighting['type'] = TypeWeighting(type_weights)
+    return weighting or None
